# Switch from case to if statements in JsonReader and code cleanup