--- a/src/System.Text.JsonLab/System/Text/Json/JsonTokenType.cs
+++ b/src/System.Text.JsonLab/System/Text/Json/JsonTokenType.cs
@@ -13,7 +13,5 @@ namespace System.Text.JsonLab
         PropertyName,
         Comment,
         Value,
-        Null,
-        Undefined,
     }
 }