@AndyAyersMS 2018-08-03
The jit's strategy for switches with small numbers of cases is something that could potentially be improved. See for instance dotnet/coreclr#12868 and dotnet/coreclr#13574.

When you convert the switch to a sequences of if tests you might also be prioritizing tests so that the overall dynamic path length is minimized based on your knowledge of likely distributions of the switch value. The jit can't do that.